refactor(frontend): moved routes of material to its own file
