+++ knotnpunkt/site/__init__.py
@@ -26,21 +26,23 @@ from ..database.db import (
     AuslagenKategorie,
 )
 from ..utils import checkverfuegbarkeit
+from .material import material_site
 
-site = Blueprint("views", __name__, template_folder="templates")
+site = Blueprint("site", __name__, template_folder="templates")
+site.register_blueprint(material_site)
 
 #testtestets
 
 @site.route('/')
 def redirectToLogin():
-    return redirect(url_for('views.login'))
+    return redirect(url_for('site.login'))
 
 
 @site.route("/login", methods=['GET', 'POST'])
 def login():
     error_msg = ""
     if current_user.is_authenticated:
-        return redirect(url_for('views.home'))
+        return redirect(url_for('site.home'))
     if request.args.get("newPassword"):
         error_msg = "Bitte melde dich mit deinem neuen Passwort an."
     if request.method == 'POST':
@@ -51,7 +53,7 @@ def login():
                 return redirect(url_for(".profil", benutzername=user.benutzername, initialLogin=True))
             elif user.check_passwort(request.form['passwort']):
                 login_user(user, remember=True)
-                return redirect(url_for('views.home'))
+                return redirect(url_for('site.home'))
             else:
                 error_msg = 'Ungültige Anmeldedaten. Bitte überprüfe Benutzername und Passwort.'
         else:
@@ -67,7 +69,7 @@ def logout():
     db.session.add(user)
     db.session.commit()
     logout_user()
-    return redirect(url_for('views.login'))
+    return redirect(url_for('site.login'))
 
 
 @site.route('/home')
@@ -149,10 +151,10 @@ def profil(benutzername):
                         db.session.add(user)
                         db.session.commit()
                         logout_user()
-                        return redirect(url_for("views.login", newPassword=True))
+                        return redirect(url_for("site.login", newPassword=True))
                     else:
                         error_msg = "Änderung fehlgeschlagen. Bitte bestätige dein Passwort."
-                        return redirect(url_for("views.profil", benutzername=benutzername, missingPwdConfirm=True))
+                        return redirect(url_for("site.profil", benutzername=benutzername, missingPwdConfirm=True))
                 db.session.add(user)
                 db.session.commit()
                 return redirect(url_for(".benutzer"))
@@ -172,145 +174,6 @@ def profil(benutzername):
         elif request.args.get('missingPwdConfirm'):
             error_msg = "Bitte bestätige das neues Passwort."
         return render_template('profil.html', user=user, roles=rollen, edit=edit_permission, hide_menu=hide_menu, error=error_msg)
-
-
-@site.route("/material", methods=['GET', 'POST'])
-@login_required
-def material():
-    if request.method == 'POST':
-        eigenschaften = {"anzahl": int(request.form.get('anzahl'))}
-        if request.form.get('beschreibung'):
-            eigenschaften['beschreibung'] = request.form.get('beschreibung')
-        if request.form.get('rhArtNummer'):
-            eigenschaften['rhArtNummer'] = request.form.get('rhArtNummer')
-        if request.form.get('verpackung'):
-            eigenschaften['verpackung'] = request.form.get('verpackung')
-        if int(eigenschaften['anzahl']) > 1:
-            eigenschaften['verfuegbar'] = 1
-        eigenschaften['zuletztGescannt'] = dt.strftime(
-            dt.now(), "%Y-%m-%d %H:%M")
-        neuesMaterial = Material(request.form.get(
-            'name'), request.form.get('kategorie'), eigenschaften)
-        db.session.add(neuesMaterial)
-        db.session.commit()
-        return redirect(url_for(".material"))
-    else:
-        materialien = Material.query.all()
-        verfuegbarkeit = checkverfuegbarkeit(materialien)
-        kategorien = Kategorie.query.all()
-        material_list = []
-        for material in materialien:
-            id = material.idMaterial
-            image = Img.query.filter_by(Material_idMaterial=id).first()
-            if image != None:
-                material_list.append(
-                    [material, base64.b64encode(image.img).decode('utf-8')])
-            else:
-                material_list.append([material, None])
-        return render_template('material.html', materialListe=material_list, kategorienListe=kategorien, verfuegbarkeit=verfuegbarkeit,  jsonRef=json, huRef=hu, dtRef=dt)
-
-
-@site.route('/material/<idMaterial>', methods=['GET'])
-@login_required
-def materialDetails(idMaterial):
-    material_details = Material.query.filter_by(idMaterial=idMaterial).all()
-    materialien = Material.query.all()
-    # Hier schon direkt Filtern ob MaterialID(Int) in Ausgeliehenem Material(Str) ist?
-    ausleihen = Ausleihe.query.order_by(desc(Ausleihe.ts_von)).all()
-    ausleihen_filtered_future = []
-    ausleihen_filtered_past = []
-    verfuegbarkeit = checkverfuegbarkeit(material_details)
-    for a in ausleihen:
-        if int(idMaterial) in [int(x) for x in a.materialien.split(",") if x.isdigit()]:
-            if a.ts_von > date.today():
-                ausleihen_filtered_future.append(a)
-            else:
-                ausleihen_filtered_past.append(a)
-    if len(ausleihen_filtered_past):
-        zuletzt_ausgeliehen_Tage = (
-            date.today() - ausleihen_filtered_past[0].ts_von).days
-    else:
-        zuletzt_ausgeliehen_Tage = None
-    kategorien = Kategorie.query.all()
-    material_images = Img.query.filter_by(Material_idMaterial=idMaterial).all()
-    img_id_list = []
-    if material_images:
-        for image in material_images:
-            img_id_list.append(
-                [image.img_id, base64.b64encode(image.img).decode('utf-8')])
-    else:
-        img_id_list.append(None)
-    return render_template('material_details.html', material_details=material_details, materialListe=materialien, kategorienListe=kategorien, ausleihListeZukunft=ausleihen_filtered_future, ausleihListeAlt=ausleihen_filtered_past, verfuegbarkeit=verfuegbarkeit, zuletzt_ausgeliehen_Tage=zuletzt_ausgeliehen_Tage, jsonRef=json, huRef=hu, dtRef=dt, images=img_id_list)
-
-
-@site.route('/material/edit/<idMaterial>', methods=['POST'])
-@login_required
-def materialDetailsEdit(idMaterial):
-    material_update = Material.query.filter_by(idMaterial=idMaterial).first()
-    material_update.name = request.form.get('name')
-    material_update.Kategorie_idKategorie = request.form.get('kategorie')
-    eigenschaften = dict(material_update.Eigenschaften)
-    debug(eigenschaften)
-    debug(f"{request.form.get('beschreibung')},{request.form.get('rhArtNummer')},{request.form.get('verpackung')}")
-    if request.form.get('beschreibung'):
-        eigenschaften['beschreibung'] = request.form.get('beschreibung')
-    if request.form.get('rhArtNummer'):
-        eigenschaften['rhArtNummer'] = request.form.get('rhArtNummer')
-    if request.form.get('verpackung'):
-        eigenschaften['verpackung'] = request.form.get('verpackung')
-    if request.form.get('anzahl'):
-        if int(request.form.get('anzahl')) > 1:
-            eigenschaften['anzahl'] = int(request.form.get('anzahl'))
-            eigenschaften['zaehlbar'] = True
-        else:
-            eigenschaften['zaehlbar'] = False
-    material_update.Eigenschaften = eigenschaften
-    debug(eigenschaften)
-    debug(material_update.Eigenschaften)
-    db.session.commit()
-    return redirect(url_for(".materialDetails", idMaterial=idMaterial))
-
-
-@site.route('/reservieren/<idMaterial>', methods=['POST'])
-@login_required
-def materialReservieren(idMaterial):
-    debug(request.form.get('reservierte_Materialien'))
-    neueReservierung = Ausleihe(
-        ersteller_benutzername=current_user.benutzername,
-        ts_erstellt=dt.now(),
-        ts_von=dt.strptime(request.form.get('reservieren_von'), "%Y-%m-%d"),
-        ts_bis=dt.strptime(request.form.get('reservieren_bis'), "%Y-%m-%d"),
-        materialien=request.form.get('reservierte_Materialien'),
-        empfaenger=request.form.get('empfaenger') if request.form.get(
-            'empfaenger') else current_user.benutzername,
-        beschreibung=request.form.get('beschreibung'))
-    db.session.add(neueReservierung)
-    db.session.commit()
-    return redirect(url_for(".material"))
-
-
-@site.route('/img/upload/<idMaterial>', methods=['POST'])
-@login_required
-def upload_img(idMaterial):
-    pic = request.files['pic']
-    if not pic:
-        return 'No pic uploaded!', 400
-    material_id = idMaterial
-    mimetype = pic.mimetype
-    if not material_id or not mimetype:
-        return 'Bad upload!', 400
-    img = Img(img=pic.read(), Material_idMaterial=material_id, mimetype=mimetype)
-    db.session.add(img)
-    db.session.commit()
-    return redirect(url_for(".materialDetails", idMaterial=idMaterial))
-
-
-@site.route('/img/delete/<id>/<idMaterial>')  # , methods=['POST']
-@login_required
-def delete_img(id, idMaterial):
-    Img.query.filter_by(img_id=id).delete()
-    db.session.commit()
-    return redirect(url_for(".materialDetails", idMaterial=idMaterial))
 
 
 @site.route("/kalender")
